refactor: Use them provider from the geti/ui/theme in the unit tests
