+++ web_ui/src/pages/annotator/tools/edit-tool/edit-polygon/edit-points.test.tsx
@@ -5,11 +5,11 @@ import '@wessberg/pointer-events';
 
 import { ReactNode } from 'react';
 
+import { ThemeProvider } from '@geti/ui/theme';
 import { fireEvent, render, screen } from '@testing-library/react';
 
 import { Polygon } from '../../../../../core/annotations/shapes.interface';
 import { ShapeType } from '../../../../../core/annotations/shapetype.enum';
-import { ThemeProvider } from '../../../../../providers/theme-provider/theme-provider.component';
 import { KeyMap } from '../../../../../shared/keyboard-events/keyboard.interface';
 import { AnnotatorContextMenuProvider } from '../../../providers/annotator-context-menu-provider/annotator-context-menu-provider.component';
 import { EditPoints } from './edit-points.component';
