+++ web_ui/src/pages/landing-page/landing-page-sidebar/organizations-picker/organizations-picker.test.tsx
@@ -1,6 +1,7 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
+import { ThemeProvider } from '@geti/ui/theme';
 import { fireEvent, screen, waitFor } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 
@@ -8,7 +9,6 @@ import { useSelectedOrganization } from '../../../../core/organizations/hook/use
 import { AccountStatus } from '../../../../core/organizations/organizations.interface';
 import { createInMemoryOnboardingService } from '../../../../core/users/services/inmemory-onboarding-service';
 import { OrganizationMetadata } from '../../../../core/users/services/onboarding-service.interface';
-import { ThemeProvider } from '../../../../providers/theme-provider/theme-provider.component';
 import { providersRender } from '../../../../test-utils/required-providers-render';
 import { OrganizationsPicker } from './organizations-picker.component';
