+++ web_ui/src/pages/project-details/components/project-models/legacy-train-model-dialog/model-templates-selection/model-template/model-template.test.tsx
@@ -1,6 +1,7 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
+import { ThemeProvider } from '@geti/ui/theme';
 import { fireEvent, render, screen, within } from '@testing-library/react';
 
 import {
@@ -8,7 +9,6 @@ import {
     PerformanceCategory,
 } from '../../../../../../../core/supported-algorithms/dtos/supported-algorithms.interface';
 import { getMockedSupportedAlgorithm } from '../../../../../../../core/supported-algorithms/services/test-utils';
-import { ThemeProvider } from '../../../../../../../providers/theme-provider/theme-provider.component';
 import { idMatchingFormat } from '../../../../../../../test-utils/id-utils';
 import { checkTooltip } from '../../../../../../../test-utils/utils';
 import { ModelTemplate } from './model-template.component';
